+++ classes/scraper.py
@@ -6,11 +6,9 @@ import os
 import tldextract
 import datetime
 import logging
-import typer
 from pathlib import Path
-from scripts.mongo import get_client, Schema
 
-class SimpleScraper():
+class Scraper():
     """
     A simple synchronous "scraper" that fetches the content of a list of pages
         and saves it to json files.
@@ -18,13 +16,14 @@ class SimpleScraper():
             scraper = SimpleScraper(['http://bdx.se','http://ssab.se'])
             scraper.scrape_all()
     """
-    def __init__(self, start_urls):
+    def __init__(self, scrape_output_folder, start_urls):
+        self.scrape_output_folder = scrape_output_folder
         self.start_urls = start_urls
 
     def _save_to_json(self, data, filename):
         logging.info('Saving scraped data to %s', filename)
-        Path('scraped_data').mkdir(parents=True, exist_ok=True)
-        full_path = os.path.join('scraped_data',filename)
+        Path(self.scrape_output_folder).mkdir(parents=True, exist_ok=True)
+        full_path = os.path.join(self.scrape_output_folder,filename)
         with open(full_path, 'w', encoding='utf-8') as f:
             json.dump(data, f, ensure_ascii=False)
 
@@ -50,20 +49,3 @@ class SimpleScraper():
             data = {'domain':domain,'url':url,'raw_html':raw}
 
             self._save_to_json(data, f"{tld_extractor.domain}_{tld_extractor.suffix}_{timestamp}.json")
-    
-def main():
-    client = get_client()
-    query = {"url": {"$regex": r"/\S"}}
-        
-    logging.debug("Querying the database with the following query: %s", query)
-    documents = client[Schema.DB][Schema.COMPANIES].find(query)
-    start_urls = [company["url"] for company in documents]
-
-    scraper = SimpleScraper(start_urls)
-    scraper.scrape_all()
-
-if __name__ == "__main__":
-    logging.basicConfig(level=logging.DEBUG)
-    typer.run(main)
-    
-
