+++ pipeline/divide_dataset.py
@@ -3,55 +3,52 @@ This script divides a dataset into a smaller dataset and a cross-validation data
 
 The smaller dataset will contain a percentage of companies with the same SNI code.
 """
-import logging
 import math
 
 import typer
 from typing_extensions import Annotated
 
-from scripts.scb import SCBinterface
-
+from adapters.train import TrainAdapter
+from adapters.extract import ExtractAdapter
+from adapters.scb import SCBAdapter
 
 def main(
-            procentage: Annotated[int, typer.Argument()] = 10
+            validation_set_size: Annotated[int, typer.Argument()] = 10
         ):
     """
-    Divide the dataset into a smaller dataset and a cross-validation dataset based on the SNI code of each company.
-    
-    
-    :param: procentage (int, optional): Procentage of the dataset that should be used for cross-validation. Defaults to 10%.
-    """
-    
-    interface = SCBinterface()
+    Divide the dataset into a smaller dataset and a validation dataset based on the SNI code of each company.
 
-    
-    
-    nr_of_each_SNI = interface.fetch_aggegrate_companies_by_sni()
+    :param: validation_set_size (int, optional): Percent of the entire dataset that should be used as a validation set. Defaults to 10%.
+    """
+
+    scb_adapter       = SCBAdapter()
+    extract_adapter   = ExtractAdapter()
+    train_adapter     = TrainAdapter()
+
+    nr_of_each_SNI = scb_adapter.aggregate_companies_by_sni()
     
     stored_sni = {}
     
     for sni in nr_of_each_SNI:
-        nr_of_cross_validation_companies = math.ceil(sni["count"] * procentage / 100)
+        nr_of_cross_validation_companies = math.ceil(sni["count"] * validation_set_size / 100)
         
         for company in sni['companies']:
-            company_scraped_data = interface.fetch_company_extracted_data(company)
+            company_scraped_data = extract_adapter.fetch_company_extracted_data(company)
             if company_scraped_data is None:
                 continue
             company_scraped_data.pop("_id")
             
-            company_data = interface.fetch_company_by_id(company)
+            company_data = scb_adapter.fetch_company_by_id(company)
             company_scraped_data["branch_codes"] = company_data["branch_codes"]
             if company_data["branch_codes"][0] not in stored_sni.keys()  or stored_sni[company_data["branch_codes"][0]] < nr_of_cross_validation_companies:
                 #Move scraped data to cross-validation dataset
-                interface.insert_to_dev_set(company_scraped_data)
+                train_adapter.insert_to_dev_set(company_scraped_data)
                 if company_data["branch_codes"][0] not in stored_sni.keys():
                     stored_sni[company_data["branch_codes"][0]] = 1
                 else:
                     stored_sni[company_data["branch_codes"][0]] += 1
             else:
-                interface.insert_to_train_set(company_scraped_data)
-                
- 
-   
+                train_adapter.insert_to_train_set(company_scraped_data)
+
 if __name__ == "__main__":
     typer.run(main)
