+++ pipeline/scb.py
@@ -1,0 +1,10 @@
+"""
+Polls SCB
+"""
+import logging
+from adapters.scb import SCBAdapter
+
+if __name__ == "__main__":
+    logging.basicConfig(level=logging.ERROR)
+    scb = SCBAdapter()
+    scb.fetch_all_companies_from_api(fetch_limit=50)
