--- scraping/NLPspider/pipelines.py
@@ -1,13 +1,0 @@
-""" Define your item pipelines here
-
-Don't forget to add your pipeline to the ITEM_PIPELINES setting
-See: https://docs.scrapy.org/en/latest/topics/item-pipeline.html
-
-
-useful for handling different item types with a single interface
-"""
-
-
-class NlpspiderPipeline:
-    def process_item(self, item, spider):
-        return item
