+++ tests/test_extract_wrapper.py
@@ -7,7 +7,7 @@ It uses the pytest library to run the tests. To run simply execute the following
 """
 import requests
 from pathlib import Path
-from scripts.extract_wrapper import extract_wrapper
+from pipeline.extract import main
 import pytest
 import tldextract
 import json
@@ -76,7 +76,7 @@ def test_structure(tmp_path, temp_scraped_data, temp_dataset):
     Test the structure of the output file generated by the extractor_wrapper function.
     """
     temp_output_file = tmp_path / "output.json"
-    extract_wrapper(temp_dataset, temp_scraped_data, temp_output_file, extract_meta=True, extract_body=False, p_only=False)
+    main(temp_dataset, temp_scraped_data, temp_output_file, extract_meta=True, extract_body=False, p_only=False)
     
     with open(temp_output_file, 'r', encoding='utf-8') as f:
         data = json.load(f)
